+++ src/main/java/com/api_salud/api_salud/service/EntidadServiceImpl.java
@@ -35,7 +35,7 @@ public class EntidadServiceImpl implements EntidadService {
     public EntidadResponse obtenerEntidadDelContexto() {
 
         // 1️⃣ Obtener el id de la entidad desde el contexto (JWT → Filter)
-        Long entidadId = TenantContext.getEntidadId();
+        Integer entidadId = TenantContext.getEntidadId();
 
         if (entidadId == null) {
             throw new RuntimeException("No se encontró la entidad en el contexto");
